combine streaming and non streaming for tencent tts
