+++ Cargo.toml
@@ -40,6 +40,7 @@ tokio-util = "0.7.16"
 tracing = "0.1.41"
 tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
 tracing-appender = "0.2"
+unic-emoji = "0.9.0"
 url = "2.5.4"
 uuid = { version = "1.18.0", features = ["v4"] }
 webrtc = "0.13.0"
